requirements.txt
PySide6>=6.7.0
pexpect>=4.9.0
PyYAML>=6.0.1
pypdf>=4.2.0
python-docx>=1.1.2
python-pptx>=0.6.23
openpyxl>=3.1.2
Pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
